requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
slowapi==0.1.9
